add some casts to suppress clang warnings
git-svn-id: https://svn.r-project.org/R/trunk@77445 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/eval.c
@@ -6536,7 +6536,7 @@ static SEXP bcEval(SEXP body, SEXP rho, Rboolean useCache)
 		for (int i = 0; i < FOR_LOOP_STATE_SIZE; i++)
 		    R_BCNodeStackTop[i] = loopdata[i];
 		R_BCNodeStackTop += FOR_LOOP_STATE_SIZE;
-		SET_FOR_LOOP_BCPROT_OFFSET(R_BCProtTop - R_BCNodeStackBase);
+		SET_FOR_LOOP_BCPROT_OFFSET((int)(R_BCProtTop - R_BCNodeStackBase));
 		INCLNK_stack(R_BCNodeStackTop);
 
 		begincontext(cntxt, CTXT_LOOP, R_NilValue, rho, R_BaseEnv,
@@ -6644,7 +6644,7 @@ static SEXP bcEval(SEXP body, SEXP rho, Rboolean useCache)
 	}
 	/* the seq, binding cell, and value on the stack are now boxed */
 
-	SET_FOR_LOOP_BCPROT_OFFSET(R_BCProtTop - R_BCNodeStackBase);
+	SET_FOR_LOOP_BCPROT_OFFSET((int)(R_BCProtTop - R_BCNodeStackBase));
 	INCLNK_stack(R_BCNodeStackTop);
 
 	BC_CHECK_SIGINT();
@@ -7624,7 +7624,7 @@ static SEXP bcEval(SEXP body, SEXP rho, Rboolean useCache)
       NEXT();
     OP(INCLNKSTK, 0):
       {
-	  int offset = R_BCProtTop - R_BCNodeStackBase;
+	  int offset = (int)(R_BCProtTop - R_BCNodeStackBase);
 	  INCLNK_stack(R_BCNodeStackTop);
 	  BCNPUSH_INTEGER(offset);
 	  NEXT();
